Move testing to subfolders to mimic structure in src
