+++ tests/properties/test_band_gap.py
