+++ tests/properties/test_energies.py
