+++ tests/properties/test_hopping_matrix.py
