+++ tests/properties/test_spectral_profile.py
